rename zukandetail class and layout
