+++ app/src/main/java/xyz/virtual_diving/projectmainver2/DB/ZukanDatabase.java
@@ -27,7 +27,6 @@ public class ZukanDatabase {
         ZukanSQLiteOpenHelper helper = new ZukanSQLiteOpenHelper(ZukanListActivity.getContext());
         mDb = helper.getWritableDatabase();
         Cursor c = mDb.query(ZukanSQLiteOpenHelper.TABLE_NAME, FROM, null, null, null, null, ORDER_BY);//queryの実行
-        if (!c.moveToPosition(id)) {
             // ContentValuesにデータを格納
             ContentValues values = new ContentValues();
             // カラム名に値を渡す
@@ -44,7 +43,7 @@ public class ZukanDatabase {
             } finally {
                 mDb.close();
             }
-        }
+
         c.close();
     }
