+++ app/src/main/res/layout/zukandetail_main.xml
@@ -1,0 +1,73 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:orientation="vertical"
+    android:layout_width="fill_parent"
+    android:layout_height="fill_parent">
+
+    <LinearLayout
+        android:orientation="horizontal"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:weightSum="1">
+        <ImageButton
+            android:src="@drawable/ic_action_back"
+            android:background="@null"
+            android:layout_width="50dp"
+            android:layout_height="50dp"
+            android:id="@+id/ZukanDetail_back"
+            android:layout_gravity="left" />
+        <View
+            android:layout_width="0dp"
+            android:layout_height="1dp"
+            android:layout_weight="1"
+            />
+        <ImageButton
+            android:src="@drawable/quiz"
+            android:background="@null"
+            android:layout_width="50dp"
+            android:layout_height="50dp"
+            android:id="@+id/ZukanDetail_quiz"
+            android:layout_gravity="right"
+             />
+    </LinearLayout>
+        <LinearLayout
+            android:orientation="horizontal"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:weightSum="1">
+            <ImageView
+                android:layout_marginLeft="50dp"
+                android:layout_width="100dp"
+                android:layout_height="100dp"
+                android:id="@+id/ZukanDetail_pic" />
+            <LinearLayout
+                android:orientation="vertical"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content" >
+                    <TextView
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:textSize="30dp"
+                        android:layout_marginLeft="50dp"
+                        android:id="@+id/ZukanDetail_fishName"/>
+                    <TextView
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:text="New Text"
+                        android:layout_marginTop="10dp"
+                        android:layout_marginLeft="50dp"
+                        android:id="@+id/ZukanDetail_abstarct" />
+            </LinearLayout>
+        </LinearLayout>
+    <View
+        android:layout_marginTop="25dp"
+        android:layout_width="match_parent"
+        android:layout_height="1px"
+        android:background="#000000" />
+    <android.support.v4.view.ViewPager
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="25dp"
+        android:id="@+id/ZukanDetail_ViewPager"/>
+</LinearLayout>
